api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Esli Heyvaert <esli.heyvaert@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
SQLAlchemy = "^1.3.22"
psycopg2-binary = "^2.8.6"
uvicorn = "^0.13.3"
python-dotenv = "^0.15.0"
alembic = "^1.4.3"
inflect = "^5.0.2"
pydantic = { extras = ["email"], version = "^1.7.3" }
passlib = "^1.7.4"
python-jose = { extras = ["cryptography"], version = "^3.2.0" }
python-multipart = "^0.0.5"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.790"
sqlalchemy-stubs = "^0.4"
flake8 = "^3.8.4"
autoflake = "^1.4"
isort = "^5.7.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
